make player movement code a bit clearer
I also responded here to the message in your comment, yusuf.
See the docs on match here for efficiently matching multiple cases and more:
https://docs.godotengine.org/en/3.1/getting_started/scripting/gdscript/gdscript_basics.html#match

By the way, I think my previous commit fixed a bug and introduced a new bug:
If player attempts to move into a boulder which is on the wrong hole shape and cannot be moved, the player
correctly does not move, but they still fall!